Clean-up cython stack for distrib. - add setup.py silliness to auto-import numpy headers. - use Py_ssize_t to fix-up int / long / size_t issues.

+++ inpoly/inpoly_.pyx
@@ -37,7 +37,7 @@ def _inpoly(np.ndarray[double, ndim=+2] vert,
 #----------------------------------- compute y-range overlap
     YMIN = node[edge[:, 0], 1] - veps
    
-    cdef np.ndarray[long] HEAD = \
+    cdef np.ndarray[Py_ssize_t] HEAD = \
         np.searchsorted(vert[:, 1], YMIN, "left" )
     
 #----------------------------------- loop over polygon edges
